+++ euporie/notebook/app.py
@@ -45,6 +45,7 @@ from euporie.core.widgets.file_browser import FileBrowser
 from euporie.core.widgets.layout import TabBarControl, TabBarTab
 from euporie.core.widgets.logo import logo_micro
 from euporie.core.widgets.menu import MenuBar, MenuItem
+from euporie.core.widgets.minimap import MiniMap
 from euporie.core.widgets.pager import Pager
 from euporie.core.widgets.palette import CommandPalette
 from euporie.core.widgets.toc import TableOfContents
@@ -261,6 +262,11 @@ class NotebookApp(BaseApp):
                         "Table of Contents",
                         "",
                         TableOfContents(),
+                    ),
+                    (
+                        "Mini Map",
+                        "",
+                        MiniMap(),
                     ),
                 ]
             ),